+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to the "jira-quick-info" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.4] 2024-08-12
+- Change extension kind to "ui" to download images next to the tooltip renderer
+- Reload PAT on change of the corresponding option
+- Added a new "Output" log tag, to print list of downloaded images and their location on disk
+- Fix default issue name from the Workspace folder basename
+
 ## [0.0.3] 2024-08-07
 - Download thumbnailed images to render them in the tooltip
 - Added a command to erase the thumbnails from the disk